// src/constants/player-position-name.enum.ts
export enum EPlayerPositionName {
	// offensive
	LCF = 'LCF',
	ST = 'ST',
	RCF = 'RCF',

	// midfield
	LWM = 'LWM',
	AML = 'AML',
	AMC = 'AMC',
	AMR = 'AMR',
	RWM = 'RWM',

	LM = 'LM',
	LCM = 'LCM',
	CM = 'CM',
	RCM = 'RCM',
	RM = 'RM',

	// defensive
	LWB = 'LWB',
	LDM = 'LDM',
	CDM = 'CDM',
	RDM = 'RDM',
	RWB = 'RWB',

	LB = 'LB',
	LCB = 'LCB',
	CB = 'CB',
	RCB = 'RCB',
	RB = 'RB',

	// GK
	GK = 'GK'
}